pyproject.toml
[project]
name = "agent-newsletter"
version = "0.1.0"
description = "Leadership Strategy Newsletter Agent - Multi-agent system for HBR-quality newsletters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LLM Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-aws>=0.2.0",

    # AWS
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # Web Research
    "tavily-python>=0.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "selenium>=4.15.0",

    # PDF Processing
    "pypdf2>=3.0.0",
    "pdfplumber>=0.10.0",

    # Data Visualization
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",

    # Image Generation
    "pillow>=10.0.0",

    # Text-to-Speech (ElevenLabs)
    "elevenlabs>=1.0.0",

    # Video Production
    "moviepy>=1.0.3",

    # Document Generation
    "weasyprint>=60.0",
    "markdown>=3.5.0",
    "jinja2>=3.1.0",

    # Quality & Validation
    "textstat>=0.7.3",

    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Observability
    "langsmith>=0.1.0",

    # Utilities
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
